joint all collapsables in one component
