+++ src/components/Cards.js
@@ -6,7 +6,9 @@ import Footer from './Footer';
 
 // const linkAdalab = 'url(https://adalab.es/)';
 
-
+const colorPallete1 = ['#104E43', '#438792', '#a2deaf'];
+const colorPallete2 = ['#420101', '#bd1010', '#e95626'];
+const colorPallete3 = ['#3e5b65', '#eab052', '#a0c0cf'];
 
 class Cards extends React.Component {
   render() {
@@ -16,7 +18,7 @@ class Cards extends React.Component {
           <Header />
           <main className="card__main">
             <CardPreview />
-            <CardForm />
+            <CardForm color={colorPallete1} color2={colorPallete2} color3={colorPallete3}/>
           </main>
           <Footer />
         </div>
