+++ src/scss/core/_class-utils.scss
@@ -79,7 +79,7 @@
 
 //common
 .hidden {
-  display: none;
+  // display: none;
 }
 
 .not-available{
